# docs/api/blueiq-api-gateway-openapi.yaml
openapi: 3.0.3
info:
  title: BlueIQ API Gateway
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
  description: |
    This document describes the API for the BlueIQ API Gateway.
    It serves as the single entry point for all client interactions with the BlueIQ microservices.
    This specification is designed to be a reference for consumers and a template for similar gateway patterns.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://dev.blueiq.example.com/api/v1
    description: Development environment
  - url: https://staging.blueiq.example.com/api/v1
    description: Staging environment
  - url: https://blueiq.example.com/api/v1
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /users:
    get:
      tags:
        - User Management
      summary: List all users
      operationId: listUsers
      description: Retrieves a paginated list of all user accounts. Supports query parameters for pagination.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of users per page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - User Management
      summary: Create a new user
      operationId: createUser
      description: Creates a new user account. Requires username, email, and password. Optional fields include first name, last name, and role assignments.
      security:
        - BearerAuth: []
      requestBody:
        description: User object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}:
    get:
      tags:
        - User Management
      summary: Get a user by ID
      operationId: getUserById
      description: Retrieves the details of a specific user account by their unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - User Management
      summary: Update an existing user
      operationId: updateUser
      description: Updates an existing user's account details. Allows modification of email, names, active status, and role assignments.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: string
            format: uuid
      requestBody:
        description: User object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - User Management
      summary: Delete a user by ID
      operationId: deleteUser
      description: Deletes a specific user account by their unique ID. This operation is typically irreversible.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid input provided.
        details:
          type: string
          nullable: true
          example: The 'username' field is required.
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
          readOnly: true
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: johndoe@example.com
        firstName:
          type: string
          example: John
          nullable: true
        lastName:
          type: string
          example: Doe
          nullable: true
        isActive:
          type: boolean
          example: true
          default: true
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["f47ac10b-58cc-4372-a567-0e02b2c3d479"]
          description: List of role IDs assigned to the user.
        createdAt:
          type: string
          format: date-time
          example: '2023-01-15T10:30:00Z'
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-16T12:45:00Z'
          readOnly: true
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        totalItems:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        currentPage:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: newuser
        email:
          type: string
          format: email
          example: newuser@example.com
        password:
          type: string
          format: password
          example: 'Str0ngP@sswOrd!'
          minLength: 8
        firstName:
          type: string
          example: New
          nullable: true
        lastName:
          type: string
          example: User
          nullable: true
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["f47ac10b-58cc-4372-a567-0e02b2c3d479"]
          description: List of role IDs to assign to the new user.
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: updateduser@example.com
        firstName:
          type: string
          example: Updated
          nullable: true
        lastName:
          type: string
          example: Userley
          nullable: true
        isActive:
          type: boolean
          example: false
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["a1b2c3d4-e5f6-7890-1234-567890abcdef"]
          description: List of role IDs to assign to the user. This will replace existing roles.
...
  responses:
    UnauthorizedError:
      description: Unauthorized - Authentication token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                statusCode: 401
                message: "Authentication token is missing."
            invalidToken:
              summary: Invalid token
              value:
                statusCode: 401
                message: "Authentication token is invalid or expired."
    ForbiddenError:
      description: Forbidden - User does not have permission to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Insufficient permissions
              value:
                statusCode: 403
                message: "You do not have the necessary permissions to perform this action."
    NotFoundError:
      description: Not Found - The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                statusCode: 404
                message: "The requested resource was not found."
                details: "User with ID '12345' does not exist."
...
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericError:
              summary: Generic server error
              value:
                statusCode: 500
                message: "An unexpected error occurred. Please try again later or contact support."
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token for authentication. The token should be included in the
        Authorization header as 'Bearer {token}'.
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`